skatlaz_1.0_alpha/requirements.txt
transformers
torch
sentence-transformers
faiss-cpu
diffusers
spacy
requests
pypdf
google-api-python-client
